# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SkyRadar REST API
  version: 1.0.0
  description: |
    SkyRadar 最终 RESTful API 契约源。所有业务接口使用 `/api/v1/*` 路径。

    成功响应统一使用 `data`、`meta`、`links`；错误响应统一使用
    `error`、`message`、`detail`、`request_id`。
servers:
  - url: /
...
paths:
...
  /api/v1/github-accounts:
    get:
      tags:
        - Settings
      summary: 获取 GitHub 账号列表
      operationId: listGithubAccounts
      responses:
        "200":
          description: GitHub 账号列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubAccountListSuccess"
    post:
      tags:
        - Settings
      summary: 新增 GitHub 账号或 token
      operationId: createGithubAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubAccountRequest"
            example:
              username: skyradar-bot
              token: <token>
      responses:
        "201":
          description: GitHub 账号已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubAccountSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/github-accounts/{username}:
    delete:
      tags:
        - Settings
      summary: 删除 GitHub 账号
      operationId: deleteGithubAccount
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          example: skyradar-bot
      responses:
        "204":
          description: GitHub 账号已删除；响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/search-rules:
    get:
      tags:
        - Settings
      summary: 获取监控规则
      operationId: listQueryRules
      responses:
        "200":
          description: 监控规则列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRuleListSuccess"
    post:
      tags:
        - Settings
      summary: 创建监控规则
      operationId: createQueryRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRuleRequest"
            example:
              keyword: '"ghp_" password'
              tag: github-token
              search_type: code
              enabled: true
              analysis_enabled: false
      responses:
        "201":
          description: 监控规则已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRuleSuccess"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/search-rules/{tag}:
    put:
      tags:
        - Settings
      summary: 更新监控规则
      operationId: updateQueryRule
      parameters:
        - $ref: "#/components/parameters/SearchRuleTag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRuleRequest"
      responses:
        "200":
          description: 监控规则已更新。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRuleSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags:
        - Settings
      summary: 删除监控规则
      operationId: deleteQueryRule
      parameters:
        - $ref: "#/components/parameters/SearchRuleTag"
      responses:
        "204":
          description: 监控规则已删除；响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/task-schedules/current:
    get:
      tags:
        - Settings
      summary: 获取当前任务调度设置
      operationId: getTaskSchedule
      responses:
        "200":
          description: 当前任务调度设置。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskScheduleSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Settings
      summary: 保存当前任务调度设置
      operationId: replaceTaskSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskScheduleRequest"
            example:
              page: 10
              minute: 30
      responses:
        "200":
          description: 当前任务调度设置已保存。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskScheduleSuccess"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/openai-settings/current:
    get:
      tags:
        - Settings
      summary: 获取 OpenAI 分析配置
      description: 响应不会返回明文 API Key，只返回是否已配置和脱敏值。
      operationId: getOpenAISetting
      responses:
        "200":
          description: OpenAI 分析配置。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAISettingSuccess"
    put:
      tags:
        - Settings
      summary: 保存 OpenAI 分析配置
      operationId: replaceOpenAISetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAISettingRequest"
            example:
              enabled: true
              api_key: sk-...
              base_url: https://api.openai.com/v1
              model: gpt-4o-mini
              prompt: 请简要分析泄露风险
              notify_webhook_on_useful: true
              usefulness_prompt: 判断项目是否符合关注方向并值得推送
              interests: 浏览器指纹与自动化规避
              max_context_lines: 120
              max_context_chars: 12000
              timeout_seconds: 30
              max_retries: 2
              concurrency: 2
      responses:
        "200":
          description: OpenAI 分析配置已保存，响应不返回明文 API Key。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAISettingSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/blacklist-items:
    get:
      tags:
        - Settings
      summary: 获取黑名单
      operationId: listBlacklistItems
      responses:
        "200":
          description: 黑名单列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistItemListSuccess"
    post:
      tags:
        - Settings
      summary: 新增黑名单项
      operationId: createBlacklistItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistItemRequest"
            example:
              text: example.com
      responses:
        "201":
          description: 黑名单项已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistItemSuccess"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/blacklist-items/{text}:
    delete:
      tags:
        - Settings
      summary: 删除黑名单项
      operationId: deleteBlacklistItem
      parameters:
        - $ref: "#/components/parameters/BlacklistText"
      responses:
        "204":
          description: 黑名单项已删除；响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/asset-rules:
    get:
      tags:
        - Settings
      summary: 获取资产提取规则
      operationId: listAssetRules
      responses:
        "200":
          description: 资产提取规则列表。未配置时返回预置 domain、email 和 ip 正则规则。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRuleListSuccess"
    post:
      tags:
        - Settings
      summary: 新增资产提取规则
      operationId: createAssetRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetRuleRequest"
            example:
              name: AWS Access Key
              type: secret
              pattern: AKIA[0-9A-Z]{16}
              enabled: true
      responses:
        "201":
          description: 资产提取规则已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRuleSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/asset-rules/{rule_id}:
    put:
      tags:
        - Settings
      summary: 更新资产提取规则
      operationId: updateAssetRule
      parameters:
        - $ref: "#/components/parameters/AssetRuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetRuleRequest"
      responses:
        "200":
          description: 资产提取规则已更新。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRuleSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags:
        - Settings
      summary: 删除资产提取规则
      operationId: deleteAssetRule
      parameters:
        - $ref: "#/components/parameters/AssetRuleId"
      responses:
        "204":
          description: 资产提取规则已删除；预置规则删除后会被隐藏并不再参与扫描，响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notification-recipients:
    get:
      tags:
        - Settings
      summary: 获取邮件通知接收人
      operationId: listNoticeRecipients
      responses:
        "200":
          description: 邮件通知接收人列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeRecipientListSuccess"
    post:
      tags:
        - Settings
      summary: 新增邮件通知接收人
      operationId: createNoticeRecipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeRecipientRequest"
            example:
              mail: security@example.com
      responses:
        "201":
          description: 邮件通知接收人已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeRecipientSuccess"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/notification-recipients/{mail}:
    delete:
      tags:
        - Settings
      summary: 删除邮件通知接收人
      operationId: deleteNoticeRecipient
      parameters:
        - name: mail
          in: path
          required: true
          schema:
            type: string
          example: security@example.com
      responses:
        "204":
          description: 邮件通知接收人已删除；响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/mail-settings/current:
    get:
      tags:
        - Settings
      summary: 获取 SMTP 配置
      operationId: getMailSetting
      responses:
        "200":
          description: SMTP 配置，不包含 SMTP 密码。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSettingSuccess"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Settings
      summary: 保存 SMTP 配置
      operationId: replaceMailSetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MailSettingRequest"
            example:
              from: skyradar@example.com
              host: smtp.example.com
              port: 587
              tls: true
              username: skyradar@example.com
              password: <password>
              domain: https://skyradar.example.com
              enabled: true
      responses:
        "200":
          description: SMTP 配置已保存，不返回 SMTP 密码。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailSettingSuccess"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/webhooks:
    get:
      tags:
        - Settings
      summary: 获取 webhook 配置
      operationId: listWebhooks
      responses:
        "200":
          description: webhook 配置列表。响应只包含脱敏 URL 和稳定 ID。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookListSuccess"
    post:
      tags:
        - Settings
      summary: 创建 webhook 配置
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
            example:
              provider: dingtalk
              webhook_url: https://oapi.dingtalk.com/robot/send?access_token=<token>
              secret: <secret>
              domain: https://skyradar.example.com
              enabled: true
      responses:
        "201":
          description: webhook 配置已创建。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/webhooks/{webhook_id}:
    delete:
      tags:
        - Settings
      summary: 删除 webhook 配置
      operationId: deleteWebhook
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
          example: webhook-001
      responses:
        "204":
          description: webhook 配置已删除；响应不返回 body。
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/webhook-tests:
    post:
      tags:
        - Settings
      summary: 同步发送 webhook 测试消息
      operationId: createWebhookTest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
      responses:
        "201":
          description: webhook 测试消息已同步投递并由 provider 确认成功。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookTestSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
components:
  parameters:
...
    SearchRuleTag:
      name: tag
      in: path
      required: true
      schema:
        type: string
      example: github-token
    BlacklistText:
      name: text
      in: path
      required: true
      schema:
        type: string
      example: obsolete-secret
    AssetRuleId:
      name: rule_id
      in: path
      required: true
      schema:
        type: string
      example: email
  responses:
    BadRequest:
      description: 请求格式或参数无法处理。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未认证或外部凭据不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 资源状态冲突。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: 请求语义校验失败。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ApiSuccess:
      type: object
      required:
        - data
      properties:
        data:
          nullable: true
        meta:
          type: object
          additionalProperties: true
        links:
          type: object
          additionalProperties:
            type: string
...
    GithubAccountListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/GithubAccount"
    GithubAccountSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/GithubAccount"
    GithubAccount:
      type: object
      required:
        - username
      properties:
        username:
          type: string
        mask_password:
          type: string
        addat:
          type: integer
        rate_limit:
          type: integer
        rate_remaining:
          type: integer
    GithubAccountRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
          writeOnly: true
    QueryRuleListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/QueryRule"
    QueryRuleSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/QueryRule"
...
    QueryRuleRequest:
      type: object
      required:
        - keyword
        - tag
      properties:
        keyword:
          type: string
        search_type:
          type: string
          enum:
            - code
            - repositories
          default: code
        tag:
          type: string
        enabled:
          type: boolean
          default: true
        analysis_enabled:
          type: boolean
          default: false
          description: 默认关闭；开启后扫描命中该规则时异步投递 AI 分析。
    OpenAISettingSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/OpenAISetting"
...
    OpenAISettingRequest:
      type: object
      properties:
        enabled:
          type: boolean
          default: false
        api_key:
          type: string
          format: password
          writeOnly: true
        base_url:
          type: string
          format: uri
        model:
          type: string
        prompt:
          type: string
        notify_webhook_on_useful:
          type: boolean
          default: false
        usefulness_prompt:
          type: string
        interests:
          type: string
        max_context_lines:
          type: integer
          minimum: 1
          maximum: 1000
        max_context_chars:
          type: integer
          minimum: 100
          maximum: 200000
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 300
        max_retries:
          type: integer
          minimum: 0
          maximum: 10
        concurrency:
          type: integer
          minimum: 1
          maximum: 10
    TaskScheduleSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TaskSchedule"
    TaskSchedule:
      type: object
      required:
        - page
        - minute
      properties:
        page:
          type: integer
          minimum: 1
        minute:
          type: integer
          minimum: 1
        pid:
          type: integer
          nullable: true
        last:
          type: integer
          nullable: true
    TaskScheduleRequest:
      type: object
      required:
        - page
        - minute
      properties:
        page:
          type: integer
          minimum: 1
        minute:
          type: integer
          minimum: 1
    BlacklistItemListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/BlacklistItem"
    BlacklistItemSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/BlacklistItem"
...
    BlacklistItemRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
    AssetRuleListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/AssetRule"
    AssetRuleSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AssetRule"
...
    AssetRuleRequest:
      type: object
      required:
        - name
        - type
        - pattern
      properties:
        name:
          type: string
        type:
          type: string
          pattern: "^[a-z][a-z0-9_-]{0,31}$"
        pattern:
          type: string
        enabled:
          type: boolean
          default: true
    NoticeRecipientListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/NoticeRecipient"
    NoticeRecipientSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/NoticeRecipient"
...
    NoticeRecipientRequest:
      type: object
      required:
        - mail
      properties:
        mail:
          type: string
          format: email
    MailSettingSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/MailSetting"
...
    MailSettingRequest:
      allOf:
        - $ref: "#/components/schemas/MailSetting"
        - type: object
          properties:
            password:
              type: string
              format: password
              writeOnly: true
    WebhookListSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Webhook"
    WebhookSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Webhook"
...
    WebhookRequest:
      type: object
      required:
        - provider
        - webhook_url
        - secret
      properties:
        provider:
          type: string
          enum:
            - dingtalk
            - feishu
        webhook_url:
          type: string
          format: uri
          writeOnly: true
        secret:
          type: string
          format: password
          writeOnly: true
        domain:
          type: string
          format: uri
        enabled:
          type: boolean
          default: true
    WebhookTestSuccess:
      allOf:
        - $ref: "#/components/schemas/ApiSuccess"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/WebhookTest"
...